plugin:twitter cache cache expiry as integer
git-svn-id: http://svn.habariproject.org/habari-extras@2306 f4a90942-1786-4aa1-b092-57c31935ddc3

+++ plugins/twitter/trunk/twitter.plugin.php
@@ -297,7 +297,7 @@ class Twitter extends Plugin
 				if (!$notices)
 					$notices[] = $notice;
 				// Cache (even errors) to avoid hitting rate limit.
-				Cache::set( 'twitter_notices', $notices, Options::get( 'twitter__cache' ), true );
+				Cache::set( 'twitter_notices', $notices, (int) Options::get( 'twitter__cache' ), true );
 			}
 		}
 		else {
